Move oss-fuzz specific worker functions to its own module.
This is in preparation for a more generalized pipeline.
