+++ docker/worker/worker.py
@@ -14,12 +14,10 @@
 # limitations under the License.
 """OSV Worker."""
 import argparse
-import datetime
 import json
 import logging
 import os
 import math
-import re
 import resource
 import shutil
 import subprocess
@@ -27,8 +25,6 @@ import sys
 import threading
 import time
 import traceback
-import tempfile
-import yaml
 
 from google.cloud import ndb
 from google.cloud import pubsub_v1
@@ -36,14 +32,12 @@ import pygit2
 
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 import osv
+import oss_fuzz
 
 DEFAULT_WORK_DIR = '/work'
 OSS_FUZZ_GIT_URL = 'https://github.com/google/oss-fuzz.git'
 TASK_SUBSCRIPTION = 'tasks'
 MAX_LEASE_DURATION = 6 * 60 * 60  # 4 hours.
-
-OSS_FUZZ_ISSUE_URL = 'https://bugs.chromium.org/p/oss-fuzz/issues/detail?id='
-OSS_FUZZ_SOURCE_PREFIX = 'oss-fuzz:'
 
 # Large projects which take way too long to build.
 # TODO(ochang): Don't hardcode this.
@@ -177,379 +171,6 @@ def clean_artifacts(oss_fuzz_dir):
     shutil.rmtree(build_dir, ignore_errors=True)
 
 
-def format_commit_range(old_commit, new_commit):
-  """Format a commit range."""
-  if old_commit == new_commit:
-    return old_commit
-
-  return (old_commit or osv.UNKNOWN_COMMIT) + ':' + new_commit
-
-
-def do_bisect(bisect_type, source_id, project_name, engine, sanitizer,
-              architecture, fuzz_target, old_commit, new_commit, testcase):
-  """Do the actual bisect."""
-  import bisector
-  import build_specified_commit
-
-  with tempfile.NamedTemporaryFile() as f:
-    f.write(testcase)
-    f.flush()
-
-    build_data = build_specified_commit.BuildData(
-        project_name=project_name,
-        engine=engine,
-        sanitizer=sanitizer,
-        architecture=architecture)
-    try:
-      result = bisector.bisect(bisect_type, old_commit, new_commit, f.name,
-                               fuzz_target, build_data)
-    except bisector.BisectError as e:
-      logging.error('Bisect failed with exception:\n%s', traceback.format_exc())
-      return bisector.Result(e.repo_url, None)
-    except Exception:
-      logging.error('Bisect failed with unexpected exception:\n%s',
-                    traceback.format_exc())
-      return None
-
-    if result.commit == old_commit:
-      logging.error('Bisect failed for testcase %s, bisected to old_commit',
-                    source_id)
-      result = None
-
-    return result
-
-
-def get_oss_fuzz_summary(crash_type, crash_state):
-  """Generate a summary from OSS-Fuzz crash type and crash state."""
-  crash_type = crash_type.splitlines()[0]
-  state_lines = crash_state.splitlines()
-  if crash_type in ('ASSERT', 'CHECK failure', 'Security CHECK failure',
-                    'Security DCHECK failure'):
-    return crash_type + ': ' + state_lines[0]
-
-  if crash_type == 'Bad-cast':
-    return state_lines[0]
-
-  if not crash_state or crash_state == 'NULL':
-    return crash_type
-
-  return crash_type + ' in ' + state_lines[0]
-
-
-def get_oss_fuzz_details(issue_id, crash_type, crash_state):
-  """Generate details from OSS-Fuzz crash type and crash state."""
-  details = ''
-  if issue_id:
-    oss_fuzz_link = OSS_FUZZ_ISSUE_URL + issue_id
-    details = f'OSS-Fuzz report: {oss_fuzz_link}\n\n'
-
-  crash_type = crash_type.replace('\n', ' ')
-  return details + (f'Crash type: {crash_type}\n'
-                    f'Crash state:\n{crash_state}')
-
-
-def get_ecosystem(oss_fuzz_dir, project_name):
-  """Get ecosystem."""
-  project_yaml_path = os.path.join(oss_fuzz_dir, 'projects', project_name,
-                                   'project.yaml')
-
-  with open(project_yaml_path) as f:
-    project_yaml = yaml.safe_load(f)
-
-  language = project_yaml.get('language', '')
-
-  ecosystems = {
-      'python': 'pypi',
-      'rust': 'cargo',
-      'go': 'golang',
-  }
-
-  # C/C++ projects from OSS-Fuzz don't belong to any package ecosystem.
-  return ecosystems.get(language, '')
-
-
-def _set_result_attributes(oss_fuzz_dir, message, entity):
-  """Set necessary fields from bisection message."""
-  project_name = message.attributes['project_name']
-  issue_id = message.attributes['issue_id'] or None
-  crash_type = message.attributes['crash_type']
-  crash_state = message.attributes['crash_state']
-  severity = message.attributes['severity'].upper()
-
-  timestamp = message.attributes['timestamp']
-  if timestamp:
-    timestamp = datetime.datetime.fromisoformat(timestamp)
-
-  entity.project = project_name
-  entity.ecosystem = get_ecosystem(oss_fuzz_dir, project_name)
-  entity.issue_id = issue_id
-  if issue_id:
-    entity.reference_urls.append(OSS_FUZZ_ISSUE_URL + issue_id)
-
-  entity.summary = get_oss_fuzz_summary(crash_type, crash_state)
-  entity.details = get_oss_fuzz_details(issue_id, crash_type, crash_state)
-
-  if severity:
-    entity.severity = severity
-
-  if timestamp:
-    entity.timestamp = timestamp
-
-
-def find_oss_fuzz_fix_via_commit(repo, start_commit, end_commit, source_id,
-                                 issue_id):
-  """Find fix commit by checking commit messages."""
-  if not source_id.startswith(OSS_FUZZ_SOURCE_PREFIX):
-    return None
-
-  # Walk through start_commit..end_commit
-  try:
-    walker = repo.walk(end_commit, pygit2.GIT_SORT_TOPOLOGICAL)
-  except KeyError:
-    logging.error('Failed to walk repo with invalid commit: %s', end_commit)
-    return None
-
-  walker.hide(start_commit)
-
-  testcase_id = source_id.split(':')[1]
-  oss_fuzz_pattern = re.compile(r'oss-?fuzz', re.IGNORECASE)
-  has_oss_fuzz_in_message = []
-  has_testcase_id_in_message = []
-  has_issue_id_in_message = []
-
-  # Look for commits with (in order of decreasing priority):
-  # - "oss-?fuzz" and the issue ID in the message.
-  # - ClusterFuzz testcase ID in the message.
-  # - "oss-?fuzz" in the message.
-
-  for commit in walker:
-    commit_message = commit.message.lower()
-    has_oss_fuzz = False
-
-    if oss_fuzz_pattern.search(commit_message):
-      has_oss_fuzz = True
-      has_oss_fuzz_in_message.append(commit)
-
-    if testcase_id in commit_message:
-      has_testcase_id_in_message.append(commit)
-
-    if issue_id and issue_id in commit_message and has_oss_fuzz:
-      has_issue_id_in_message.append(commit)
-
-  if has_issue_id_in_message:
-    return str(has_issue_id_in_message[0].id)
-
-  if has_testcase_id_in_message:
-    return str(has_testcase_id_in_message[0].id)
-
-  if len(has_oss_fuzz_in_message) == 1:
-    # Only pick the commit if there is a single one that mentions oss-fuzz.
-    return str(has_oss_fuzz_in_message[0].id)
-
-  return None
-
-
-def process_bisect_task(oss_fuzz_dir, bisect_type, source_id, message):
-  """Process a bisect task."""
-  bisect_type = message.attributes['type']
-  project_name = message.attributes['project_name']
-  engine = 'libfuzzer'
-  architecture = message.attributes['architecture'] or 'x86_64'
-  sanitizer = message.attributes['sanitizer']
-  fuzz_target = message.attributes['fuzz_target']
-  old_commit = message.attributes['old_commit']
-
-  new_commit = message.attributes['new_commit']
-  testcase = message.data
-  logging.info(
-      'Performing %s bisect on source_id=%s, project=%s, engine=%s, '
-      'architecture=%s, sanitizer=%s, fuzz_target=%s, old_commit=%s, '
-      'new_commit=%s', bisect_type, source_id, project_name, engine,
-      architecture, sanitizer, fuzz_target, old_commit, new_commit)
-
-  result = None
-  if project_name in PROJECT_DENYLIST:
-    logging.info('Skipping bisect for denylisted project %s', project_name)
-  elif not old_commit:
-    logging.info('Skipping bisect since there is no old_commit.')
-  else:
-    result = do_bisect(bisect_type, source_id, project_name, engine, sanitizer,
-                       architecture, fuzz_target, old_commit, new_commit,
-                       testcase)
-
-  if result.repo_url in REPO_DENYLIST:
-    logging.info('Skipping because of denylisted repo %s.', result.repo_url)
-    return
-
-  if bisect_type == 'fixed':
-    entity = osv.FixResult(id=source_id)
-  else:
-    assert bisect_type == 'regressed'
-    entity = osv.RegressResult(id=source_id)
-
-  _set_result_attributes(oss_fuzz_dir, message, entity)
-
-  if result and result.commit:
-    logging.info('Bisected to %s', result.commit)
-    entity.commit = result.commit
-    entity.repo_url = result.repo_url
-  else:
-    logging.info(
-        'Bisect not successfully performed. Setting commit range from request.')
-    entity.commit = format_commit_range(old_commit, new_commit)
-    entity.repo_url = result.repo_url if result else None
-    entity.error = 'Bisect error'
-
-  entity.put()
-
-
-def update_affected_commits(bug_id, result, project, ecosystem, public):
-  """Update affected commits."""
-  to_put = []
-  to_delete = []
-
-  for commit in result.commits:
-    affected_commit = osv.AffectedCommit(
-        id=bug_id + '-' + commit,
-        bug_id=bug_id,
-        commit=commit,
-        confidence=result.confidence,
-        project=project,
-        ecosystem=ecosystem,
-        public=public)
-
-    to_put.append(affected_commit)
-
-  # Delete any affected commits that no longer apply. This can happen in cases
-  # where a FixResult comes in later and we had previously marked a commit prior
-  # to the fix commit as being affected by a vulnerability.
-  for existing in osv.AffectedCommit.query(osv.AffectedCommit.bug_id == bug_id):
-    if existing.commit not in result.commits:
-      to_delete.append(existing.key)
-
-  ndb.put_multi(to_put)
-  ndb.delete_multi(to_delete)
-
-
-def process_impact_task(source_id, message):
-  """Process an impact task."""
-  logging.info('Processing impact task for %s', source_id)
-
-  regress_result = ndb.Key(osv.RegressResult, source_id).get()
-  if not regress_result:
-    logging.error('Missing RegressResult for %s', source_id)
-    return
-
-  fix_result = ndb.Key(osv.FixResult, source_id).get()
-  if not fix_result:
-    logging.warning('Missing FixResult for %s', source_id)
-    fix_result = osv.FixResult()
-
-  # Check if there is an existing Bug for the same source, but with a different
-  # allocated ID. This shouldn't happen.
-  allocated_bug_id = message.attributes['allocated_id']
-
-  existing_bug = osv.Bug.query(osv.Bug.source_id == source_id).get()
-  if existing_bug and existing_bug.key.id() != allocated_bug_id:
-    logging.error('Bug entry already exists for %s with a different ID %s',
-                  source_id, existing_bug.key.id())
-    return
-
-  if existing_bug and existing_bug.status == osv.BugStatus.INVALID:
-    logging.warning('Bug %s already marked as invalid.', existing_bug.key.id())
-    return
-
-  if existing_bug:
-    public = existing_bug.public
-  else:
-    raise osv.ImpactError('Task requested without Bug allocated.')
-
-  # TODO(ochang): Handle changing repo types? e.g. SVN -> Git.
-
-  repo_url = regress_result.repo_url or fix_result.repo_url
-  if not repo_url:
-    raise osv.ImpactError('No repo_url set')
-
-  issue_id = fix_result.issue_id or regress_result.issue_id
-  fix_commit = fix_result.commit
-
-  with tempfile.TemporaryDirectory() as tmp_dir:
-    repo = osv.clone_with_retries(repo_url, tmp_dir)
-
-    # If not a precise fix commit, try to find the exact one by going through
-    # commit messages (oss-fuzz only).
-    if source_id.startswith(OSS_FUZZ_SOURCE_PREFIX) and ':' in fix_commit:
-      start_commit, end_commit = fix_commit.split(':')
-      commit = find_oss_fuzz_fix_via_commit(repo, start_commit, end_commit,
-                                            source_id, issue_id)
-      if commit:
-        logging.info('Found exact fix commit %s via commit message (oss-fuzz)',
-                     commit)
-        fix_commit = commit
-
-    # Actually compute the affected commits/tags.
-    result = osv.get_affected(repo, regress_result.commit, fix_commit)
-    logging.info('Found affected %s', ', '.join(result.tags))
-
-  # If the range resolved to a single commit, simplify it.
-  if len(result.fix_commits) == 1:
-    fix_commit = result.fix_commits[0]
-  elif not result.fix_commits:
-    # Not fixed.
-    fix_commit = ''
-
-  if len(result.regress_commits) == 1:
-    regress_commit = result.regress_commits[0]
-  else:
-    regress_commit = regress_result.commit
-
-  project = fix_result.project or regress_result.project
-  ecosystem = fix_result.ecosystem or regress_result.ecosystem
-  summary = fix_result.summary or regress_result.summary
-  details = fix_result.details or regress_result.details
-  severity = fix_result.severity or regress_result.severity
-  reference_urls = fix_result.reference_urls or regress_result.reference_urls
-
-  update_affected_commits(allocated_bug_id, result, project, ecosystem, public)
-
-  existing_bug.repo_url = repo_url
-  existing_bug.fixed = fix_commit
-  existing_bug.regressed = regress_commit
-  existing_bug.affected = result.tags
-  existing_bug.affected_fuzzy = osv.normalize_tags(result.tags)
-  existing_bug.confidence = result.confidence
-  existing_bug.issue_id = issue_id
-  existing_bug.project = project
-  existing_bug.ecosystem = ecosystem
-  existing_bug.summary = summary
-  existing_bug.details = details
-  existing_bug.status = osv.BugStatus.PROCESSED
-  existing_bug.severity = severity
-  existing_bug.reference_urls = reference_urls
-
-  existing_bug.additional_commit_ranges = []
-  # Don't display additional ranges for imprecise commits, as they can be
-  # confusing.
-  if ':' in existing_bug.fixed or ':' in existing_bug.regressed:
-    existing_bug.put()
-    return
-
-  def _sort_key(value):
-    # Allow sorting of None values.
-    return (value[0] or '', value[1] or '')
-
-  for introduced_in, fixed_in in sorted(result.affected_ranges, key=_sort_key):
-    if (introduced_in == existing_bug.regressed and
-        (fixed_in or '') == existing_bug.fixed):
-      # Don't include the main range.
-      continue
-
-    existing_bug.additional_commit_ranges.append(
-        osv.CommitRange(introduced_in=introduced_in, fixed_in=fixed_in))
-
-  existing_bug.put()
-
-
 def find_bugs_for_tag(project_name, tag, public):
   """Find bugs for a given project and tag."""
   query = osv.Bug.query(osv.Bug.project == project_name,
@@ -608,118 +229,107 @@ def get_source_id(message):
 
   testcase_id = message.attributes['testcase_id']
   if testcase_id:
-    return OSS_FUZZ_SOURCE_PREFIX + testcase_id
+    return oss_fuzz.SOURCE_PREFIX + testcase_id
 
   return None
 
 
-def do_process_task(oss_fuzz_dir, subscriber, subscription, ack_id, message,
-                    done_event):
-  """Process task with timeout."""
-  try:
-    with _ndb_client.context():
-      source_id = get_source_id(message)
-      _state.source_id = source_id
+class TaskRunner:
+  """Task runner."""
 
-      task_type = message.attributes['type']
-      if task_type in ('regressed', 'fixed'):
-        process_bisect_task(oss_fuzz_dir, task_type, source_id, message)
-      elif task_type == 'impact':
-        try:
-          process_impact_task(source_id, message)
-        except osv.ImpactError:
-          logging.error('Failed to process impact: %s', traceback.format_exc())
-      elif task_type == 'package_info':
-        process_package_info_task(message)
-      elif task_type == 'invalid':
-        mark_bug_invalid(message)
+  def __init__(self, oss_fuzz_dir):
+    self._oss_fuzz_dir = oss_fuzz_dir
 
-      _state.source_id = None
-      subscriber.acknowledge(subscription=subscription, ack_ids=[ack_id])
-  except Exception:
-    logging.error('Unexpected exception while processing task: %s',
-                  traceback.format_exc())
-    subscriber.modify_ack_deadline(
-        subscription=subscription, ack_ids=[ack_id], ack_deadline_seconds=0)
-  finally:
-    logging.info('Ending task')
-    done_event.set()
-
-
-def handle_timeout(subscriber, subscription, ack_id, oss_fuzz_dir, message):
-  """Handle a timeout."""
-  subscriber.acknowledge(subscription=subscription, ack_ids=[ack_id])
-
-  bisect_type = message.attributes['type']
-  source_id = get_source_id(message)
-
-  logging.error('Task %s timed out (source_id=%s)', bisect_type, source_id)
-
-  if bisect_type not in ('fixed', 'regressed'):
-    return
-
-  old_commit = message.attributes['old_commit']
-  new_commit = message.attributes['new_commit']
-
-  if bisect_type == 'fixed':
-    entity = osv.FixResult(id=source_id)
-  else:
-    assert bisect_type == 'regressed'
-    entity = osv.RegressResult(id=source_id)
-
-  _set_result_attributes(oss_fuzz_dir, message, entity)
-
-  entity.commit = format_commit_range(old_commit, new_commit)
-  entity.error = 'Timeout'
-  entity.put()
-
-
-def task_loop(oss_fuzz_dir):
-  """Task loop."""
-
-  subscriber = pubsub_v1.SubscriberClient()
-
-  cloud_project = os.environ['GOOGLE_CLOUD_PROJECT']
-  subscription = subscriber.subscription_path(cloud_project, TASK_SUBSCRIPTION)
-
-  def process_task(ack_id, message):
-    """Process a task."""
-    ensure_updated_checkout(OSS_FUZZ_GIT_URL, oss_fuzz_dir)
-    clean_artifacts(oss_fuzz_dir)
-
-    # Enforce timeout by doing the work in another thread.
-    done_event = threading.Event()
-    thread = threading.Thread(
-        target=do_process_task,
-        args=(oss_fuzz_dir, subscriber, subscription, ack_id, message,
-              done_event),
-        daemon=True)
-    thread.start()
-
-    done = done_event.wait(timeout=MAX_LEASE_DURATION)
-    logging.info('Returned from task thread')
-    if not done:
-      handle_timeout(subscriber, subscription, ack_id, oss_fuzz_dir, message)
-      logging.error('Timed out processing task')
-
-  while True:
-    response = subscriber.pull(subscription=subscription, max_messages=1)
-    if not response.received_messages:
-      continue
-
-    message = response.received_messages[0].message
-    ack_id = response.received_messages[0].ack_id
-
-    leaser_done = threading.Event()
-    leaser = _PubSubLeaserThread(subscriber, subscription, ack_id, leaser_done,
-                                 MAX_LEASE_DURATION)
-    leaser.start()
-
+  def _do_process_task(self, subscriber, subscription, ack_id, message,
+                       done_event):
+    """Process task with timeout."""
     try:
-      process_task(ack_id, message)
+      with _ndb_client.context():
+        source_id = get_source_id(message)
+        _state.source_id = source_id
+
+        task_type = message.attributes['type']
+        if task_type in ('regressed', 'fixed'):
+          oss_fuzz.process_bisect_task(self._oss_fuzz_dir, task_type, source_id,
+                                       message)
+        elif task_type == 'impact':
+          try:
+            oss_fuzz.process_impact_task(source_id, message)
+          except osv.ImpactError:
+            logging.error('Failed to process impact: %s',
+                          traceback.format_exc())
+        elif task_type == 'package_info':
+          process_package_info_task(message)
+        elif task_type == 'invalid':
+          mark_bug_invalid(message)
+
+        _state.source_id = None
+        subscriber.acknowledge(subscription=subscription, ack_ids=[ack_id])
+    except Exception:
+      logging.error('Unexpected exception while processing task: %s',
+                    traceback.format_exc())
+      subscriber.modify_ack_deadline(
+          subscription=subscription, ack_ids=[ack_id], ack_deadline_seconds=0)
     finally:
-      leaser_done.set()
-    leaser.join()
+      logging.info('Ending task')
+      done_event.set()
+
+  def handle_timeout(self, subscriber, subscription, ack_id, message):
+    """Handle a timeout."""
+    subscriber.acknowledge(subscription=subscription, ack_ids=[ack_id])
+    task_type = message.attributes['type']
+    source_id = get_source_id(message)
+
+    logging.error('Task %s timed out (source_id=%s)', task_type, source_id)
+    if task_type in ('fixed', 'regressed'):
+      oss_fuzz.handle_timeout(task_type, source_id, self._oss_fuzz_dir, message)
+
+  def loop(self):
+    """Task loop."""
+    subscriber = pubsub_v1.SubscriberClient()
+
+    cloud_project = os.environ['GOOGLE_CLOUD_PROJECT']
+    subscription = subscriber.subscription_path(cloud_project,
+                                                TASK_SUBSCRIPTION)
+
+    def process_task(ack_id, message):
+      """Process a task."""
+      ensure_updated_checkout(OSS_FUZZ_GIT_URL, self._oss_fuzz_dir)
+      clean_artifacts(self._oss_fuzz_dir)
+
+      # Enforce timeout by doing the work in another thread.
+      done_event = threading.Event()
+      thread = threading.Thread(
+          target=self._do_process_task,
+          args=(self._oss_fuzz_dir, subscriber, subscription, ack_id, message,
+                done_event),
+          daemon=True)
+      thread.start()
+
+      done = done_event.wait(timeout=MAX_LEASE_DURATION)
+      logging.info('Returned from task thread')
+      if not done:
+        self.handle_timeout(subscriber, subscription, ack_id, message)
+        logging.error('Timed out processing task')
+
+    while True:
+      response = subscriber.pull(subscription=subscription, max_messages=1)
+      if not response.received_messages:
+        continue
+
+      message = response.received_messages[0].message
+      ack_id = response.received_messages[0].ack_id
+
+      leaser_done = threading.Event()
+      leaser = _PubSubLeaserThread(subscriber, subscription, ack_id,
+                                   leaser_done, MAX_LEASE_DURATION)
+      leaser.start()
+
+      try:
+        process_task(ack_id, message)
+      finally:
+        leaser_done.set()
+      leaser.join()
 
 
 def main():
@@ -751,7 +361,8 @@ def main():
   sys.path.append(os.path.join(oss_fuzz_dir, 'infra'))
 
   ensure_updated_checkout(OSS_FUZZ_GIT_URL, oss_fuzz_dir)
-  task_loop(oss_fuzz_dir)
+  task_runner = TaskRunner(oss_fuzz_dir)
+  task_runner.loop()
 
 
 if __name__ == '__main__':
